(TK-316) Add metrics support
This commit adds metrics support to the http client (clojure and java, sync
and async). A metric registry can optionally be passed into the client on
creation. If a metric registry is present, a metric will be added to time each
request. The metric name is currently `puppetlabs.http-client.<url>.<HTTP
method>` but in the future (for TK-317) there will be a metric-id request
option.

This commit also adds a `get-client-metrics`/`.getClientMetrics` function that
takes a client instance and returns the http client-specific metrics from the
metric registry.