What?
This commit adds a `get-client-metrics-data` function to return a summary of
the metrics data for each metric-id the client has registered.

However, in adding tests for this, I found that the metrics are not working as
expected.

Specifically, if I create some routes and add a `Thread/sleep` to them, this
sleep is not reflected in the time the metrics returned by the requests to
these routes. Not only do the relative amount of times for each request not
make sense, the total amount of time doesn't even reach a millisecond, despite
the fact the Thread/sleep's are in milliseconds.